+++ CHANGELOG.md
@@ -7,11 +7,14 @@
 * use `json-bigint` to parse responses, fixes [#52][i52]. This adds support for displaying numbers bigger
   than `9007199254740991`. I still recommend to use strings for numbers like this if you do not need the value for
   calculations.
+* add import/export for elasticvue settings, fixes [#53][i53]
 * improve visibility of active buttons in code viewer/editor
 * autofocus code editor in modals
 * adds a link to an upcoming survey about elasticvue
 
 [i52]: https://github.com/cars10/elasticvue/issues/52
+
+[i53]: https://github.com/cars10/elasticvue/issues/53
 
 ## 0.33.0